backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
scikit-learn
PyPDF2
spacy
pydantic
requests
psutil
# en-core-web-md
faiss-cpu
langchain
langchain-openai
langchain-community
python-docx
faiss-cpu
transformers
torch
sentence-transformers
